+++ src/textual/demo.py
@@ -15,7 +15,7 @@ from rich.text import Text
 from textual.app import App, ComposeResult
 from textual.binding import Binding
 from textual.containers import Container, Horizontal
-from textual.reactive import reactive, watch
+from textual.reactive import reactive
 from textual.widgets import (
     Button,
     Checkbox,
@@ -203,7 +203,7 @@ class DarkSwitch(Horizontal):
         yield Static("Dark mode toggle", classes="label")
 
     def on_mount(self) -> None:
-        watch(self.app, "dark", self.on_dark_change, init=False)
+        self.watch(self.app, "dark", self.on_dark_change, init=False)
 
     def on_dark_change(self, dark: bool) -> None:
         self.query_one(Checkbox).value = self.app.dark