+++ src/textual/widgets/_header.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from rich.text import Text
 
 from ..widget import Widget
-from ..reactive import Reactive, watch
+from ..reactive import Reactive
 
 
 class HeaderIcon(Widget):
@@ -133,5 +133,5 @@ class Header(Widget):
         def set_sub_title(sub_title: str) -> None:
             self.query_one(HeaderTitle).sub_text = sub_title
 
-        watch(self.app, "title", set_title)
-        watch(self.app, "sub_title", set_sub_title)
+        self.watch(self.app, "title", set_title)
+        self.watch(self.app, "sub_title", set_sub_title)